Set up auto modes and package structure
base readme too!
