+++ src/org/usfirst/frc/team2706/robot/commands/autonomous/core/StraightDriveWithEncoders.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team2706.robot.commands.autonomous.movements;
+package org.usfirst.frc.team2706.robot.commands.autonomous.core;
 
 import org.usfirst.frc.team2706.robot.Log;
 import org.usfirst.frc.team2706.robot.Robot;
@@ -7,29 +7,31 @@ import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
- * Drives the robot in a straight line towards the target found by the camera. Used for lining up
- * the peg at short distances
+ * Have the robot drive certain distance
  */
-public class StraightDriveWithCamera extends Command {
+public class StraightDriveWithEncoders extends Command {
 
-    private double speed;
+    private final double speed;
 
     private final double distance;
 
     private final double error;
 
-    private final PIDController PID;
+    private PIDController PID;
 
-    private final double P = 0.5, I = 0.06, D = 0.25;
+    private final int minDoneCycles;
+
+    private final double P = 7.5, I = 2.0, D = 25, F = 0;
 
     /**
-     * Drive at a specific speed based on camera
+     * Drive at a specific speed for a certain amount of time
      * 
      * @param speed Speed in range [-1,1]
      * @param distance The encoder distance to travel
-     * @param error The range that the robot is happy ending the command in inches
+     * @param error The range that the robot is happy ending the command in
      */
-    public StraightDriveWithCamera(double speed, double distance, double error) {
+    public StraightDriveWithEncoders(double speed, double distance, double error,
+                    int minDoneCycles) {
         requires(Robot.driveTrain);
 
         this.speed = speed;
@@ -38,13 +40,16 @@ public class StraightDriveWithCamera extends Command {
 
         this.error = error / 12.0;
 
-        PID = new PIDController(P, I, D, Robot.driveTrain.getDistanceSensorPIDSource(),
-                        Robot.driveTrain.getDrivePIDOutput(true, true, true));
+        this.minDoneCycles = minDoneCycles;
+
+        PID = new PIDController(P, I, D, F, Robot.driveTrain.getAverageEncoderPIDSource(),
+                        Robot.driveTrain.getDrivePIDOutput(true, false, false));
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
-        Robot.driveTrain.resetEncoders();
+        Log.d("StraightDrive", "Initialize");
+        Robot.driveTrain.reset();
 
         Robot.driveTrain.brakeMode(true);
 
@@ -62,30 +67,37 @@ public class StraightDriveWithCamera extends Command {
 
         PID.setSetpoint(distance);
 
+
         // Will accept within 5 inch of target
         PID.setAbsoluteTolerance(error);
-        Log.d("StraightDriveWithCamera", "init");
+
         // Start going to location
         PID.enable();
     }
 
-    protected void execute() {
-        Robot.camera.GetTargets(true);
-    }
+    private int doneTicks;
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-        return Robot.driveTrain.getDistanceToObstacle() < distance;
+        if (PID.onTarget())
+            doneTicks++;
+        else
+            doneTicks = 0;
+
+        return doneTicks >= minDoneCycles;
     }
 
     // Called once after isFinished returns true
     protected void end() {
-        // Robot.camera.enableRingLight(false);
+        Log.d("StraightDrive", "ending");
         Robot.driveTrain.brakeMode(false);
-
+        // Robot.driveTrain.brakeMode(false);
+        
         // Disable PID output and stop robot to be safe
         PID.disable();
         Robot.driveTrain.drive(0, 0);
+
+        Log.d("StraightDriveWithEncoders", "Ended");
     }
 
     // Called when another command which requires one or more of the same
