+++ src/org/usfirst/frc/team2706/robot/commands/autonomous/experimential/curvedrive/CubicEquation.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team2706.robot.commands.autonomous.curvecreator;
+package org.usfirst.frc.team2706.robot.commands.autonomous.experimential.curvedrive;
 
 import java.text.DecimalFormat;
