+++ src/org/usfirst/frc/team2706/robot/commands/autonomous/experimential/curvedrive/CurveDrive.java
@@ -1,8 +1,6 @@
-package org.usfirst.frc.team2706.robot.commands.autonomous.movements;
+package org.usfirst.frc.team2706.robot.commands.autonomous.experimential.curvedrive;
 
 import org.usfirst.frc.team2706.robot.Robot;
-import org.usfirst.frc.team2706.robot.commands.autonomous.curvecreator.CubicEquation;
-import org.usfirst.frc.team2706.robot.commands.autonomous.curvecreator.EquationCreator;
 
 import edu.wpi.first.wpilibj.command.Command;
