+++ src/org/usfirst/frc/team2706/robot/commands/autonomous/experimential/recordreplay/RecordJoystick.java
@@ -1,11 +1,11 @@
-package org.usfirst.frc.team2706.robot.commands.teleop;
+package org.usfirst.frc.team2706.robot.commands.autonomous.experimential.recordreplay;
 
 import java.util.function.Supplier;
 
 import org.usfirst.frc.team2706.robot.Log;
 import org.usfirst.frc.team2706.robot.OI;
 import org.usfirst.frc.team2706.robot.Robot;
-import org.usfirst.frc.team2706.robot.controls.RecordableJoystick;
+import org.usfirst.frc.team2706.robot.commands.teleop.ArcadeDriveWithJoystick;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.command.Command;
